bring back debug as a module
